action buttons plugin: limiting the size of action buttons.
Action buttons size is set to max panel_size / nrows.
Previously buttons were scaled up to the whole panel size,
which was inconsistent with other plugins (app menu, launchers etc).


Button orientation is not changed to minimize difference from
rev. 4.8. Perhaps the layout mechanism should be reworked anyway.

It would also be nice to have action buttons arranged horizontally
by default in the deskbar mode.

I've tried that but it collided with the invert-orientation property.
If the user set invert-orientation=TRUE to have action buttons
laid out perpendicularly in the horizontal|vertical mode, they would
become parallel to the panel in the deskbar mode, which is probably
not what the user intended.