systray plugin: limiting icon size to row size, fixing orientation.
Icon size is now max row size.
The user controlled max icon size is still available for 4.8 compat.
Maybe a better way would be to allow the user to set
a min number of rows instead of max icon size?

Centering of group of icons in the panel is disabled for consistency
with other plugins and the itembar behavior. (icons are flushed to
the top (horizontal mode) or to the left).

Allocation order of icons is changed to mimic that of itembar.
Icons are allocated along the columns (not rows, as it was before).

Haven't tested the support for non-square icons
- hopefully it still works. Can anyone confirm that?