panel-itembar: Removed item size checks.
Problem:

get_requisition on external plugins returns an old size_request
value when the size of the panel is changed quickly (e.g. when
user controls the "size" slider using cursor keys).

In case of "small" external plugins this means that they are
"losing" their "small" status when the user quickly decreases
the panel size. This is because their get_requisition size
is temporarily larger than the row size. This in turn causes
a lot of flicker when the user changes (decreases) the panel size.

(the above was tested using an xfce4-mixer plugin with a patch from:
https://bugzilla.xfce.org/show_bug.cgi?id=8350 )

Arguably, the itembar doesn't have to check the plugin size, only
the "small" flag. It's then the plugins responsibility to render itself
in such a way it fits in a single row.

The itembar is still using the plugin size to decide the row height.
It may temporarily be incorrect when the panel size changes quickly
but that doesn't cause flicker.