systray: fixing visual artifacts, bug #8399.
cols variable (number of columns in horizontal mode) was calculated
incorrectly and wasn't checked during allocation.

At start-up (in vertical modes) the allocation->height was occasionally
equal to 1, which made some icons to be temporarily allocated outside
the systray. Some applications didn't like it.

In addition to the actual fix, an optimization "fast path" was added to
filter out common start-up cases early.

A similar (although not that often) issue was observed in older systray
versions, which could indicate an error in coordinate checking. Coordinate
checking is now replaced with slot column/row checking.