Fixing autohide (hopefully correctly this time)
In previous commit on-screen panel position was always [0,x],
which didn't work well with partial struts and vertical panel on the right.

The fix uses a real panel position as returned by p_w_s_allocate_set_xy.
The only change to this functions are NULL pointer checks for efficiency
- we only need x-coord here, and for safety.

Also, a 1px adjustment in the collision checking macro.