Item-bar don't call size-request twice and fix DnD highlighting.
Only one of the item-bar dimensions has to be calculated by iterating
through all plugins. The other is just the size (width) of the panel.

Next step would be to further optimize size-request by changing it to
"preferred-length" and make use of Gtk3's minimum-size constraints.

Also, fixed the location of DnD markers.